# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "browserhand",
    "description": "A reliable, deterministic web-automation API for AI agents. Managed headless-browser sessions with selector-free perception (accessibility snapshot + stable refs, page content, screenshots) and structured action primitives. No server-side LLM — the calling agent reasons; browserhand executes. Each endpoint maps cleanly to a single tool.",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Set only if BROWSERHAND_API_KEY is configured."
      }
    },
...
  },
  "paths": {
    "/v1/sessions": {
      "post": {
        "summary": "Create a browser session",
        "tags": [
          "Sessions"
        ],
        "description": "Opens an isolated headless browser session and returns its id plus a CDP connect URL.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "description": "Options for a new browser session. All fields optional.",
                "properties": {
                  "viewport": {
                    "type": "object",
                    "additionalProperties": false,
                    "properties": {
                      "width": {
                        "type": "integer",
                        "minimum": 200,
                        "maximum": 3840
                      },
                      "height": {
                        "type": "integer",
                        "minimum": 200,
                        "maximum": 2160
                      }
                    },
                    "required": [
                      "width",
                      "height"
                    ],
                    "description": "Browser viewport size. Defaults to 1280x800."
                  },
                  "timeoutMs": {
                    "type": "integer",
                    "minimum": 1000,
                    "description": "Idle timeout for this session in ms. Defaults to the server setting."
                  },
                  "context": {
                    "type": "object",
                    "additionalProperties": true,
                    "description": "Portable state from GET /v1/sessions/:id/context, to reuse cookies/localStorage (e.g. a logged-in session)."
                  }
                }
              }
            }
          },
          "description": "Options for a new browser session. All fields optional."
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "429": {
            "description": "Structured error. `code` is stable and machine-readable; `message` states how to recover.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Structured error. `code` is stable and machine-readable; `message` states how to recover.",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Stable error code, e.g. \"stale_ref\", \"session_not_found\"."
                        },
                        "message": {
                          "type": "string",
                          "description": "Human/agent-readable explanation and recovery hint."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "503": {
            "description": "Structured error. `code` is stable and machine-readable; `message` states how to recover.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Structured error. `code` is stable and machine-readable; `message` states how to recover.",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Stable error code, e.g. \"stale_ref\", \"session_not_found\"."
                        },
                        "message": {
                          "type": "string",
                          "description": "Human/agent-readable explanation and recovery hint."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List open sessions",
        "tags": [
          "Sessions"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
    "/v1/sessions/{id}": {
      "get": {
        "summary": "Get session status",
        "tags": [
          "Sessions"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Session id from POST /v1/sessions."
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "404": {
            "description": "Structured error. `code` is stable and machine-readable; `message` states how to recover.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Structured error. `code` is stable and machine-readable; `message` states how to recover.",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Stable error code, e.g. \"stale_ref\", \"session_not_found\"."
                        },
                        "message": {
                          "type": "string",
                          "description": "Human/agent-readable explanation and recovery hint."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Close a session",
        "tags": [
          "Sessions"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Session id from POST /v1/sessions."
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "404": {
            "description": "Structured error. `code` is stable and machine-readable; `message` states how to recover.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Structured error. `code` is stable and machine-readable; `message` states how to recover.",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Stable error code, e.g. \"stale_ref\", \"session_not_found\"."
                        },
                        "message": {
                          "type": "string",
                          "description": "Human/agent-readable explanation and recovery hint."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/sessions/{id}/snapshot": {
      "get": {
        "summary": "Accessibility snapshot with stable element refs",
        "tags": [
          "Perception"
        ],
        "description": "Returns the interactive/structural elements on the page, each with a `ref` used by the action endpoints, plus a readable outline. This is the selector-free basis for acting.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Session id from POST /v1/sessions."
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "404": {
            "description": "Structured error. `code` is stable and machine-readable; `message` states how to recover.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Structured error. `code` is stable and machine-readable; `message` states how to recover.",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Stable error code, e.g. \"stale_ref\", \"session_not_found\"."
                        },
                        "message": {
                          "type": "string",
                          "description": "Human/agent-readable explanation and recovery hint."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/sessions/{id}/content": {
      "get": {
        "summary": "Readable page text",
        "tags": [
          "Perception"
        ],
        "description": "Returns the visible text of the page for the agent to read/extract. `truncated` signals the text was clipped.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Session id from POST /v1/sessions."
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "404": {
            "description": "Structured error. `code` is stable and machine-readable; `message` states how to recover.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Structured error. `code` is stable and machine-readable; `message` states how to recover.",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Stable error code, e.g. \"stale_ref\", \"session_not_found\"."
                        },
                        "message": {
                          "type": "string",
                          "description": "Human/agent-readable explanation and recovery hint."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/sessions/{id}/screenshot": {
      "get": {
        "summary": "PNG screenshot",
        "tags": [
          "Perception"
        ],
        "description": "Returns image/png so the agent can see the page. Use ?fullPage=true for the whole page or ?ref=eN for one element.",
        "parameters": [
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "fullPage",
            "required": false,
            "description": "Capture the full scrollable page. Default false (viewport only)."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "ref",
            "required": false,
            "description": "Optional: screenshot just this element."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Session id from POST /v1/sessions."
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/sessions/{id}/navigate": {
      "post": {
        "summary": "Navigate to a URL",
        "tags": [
          "Actions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "url": {
                    "type": "string",
                    "description": "Absolute URL to navigate to (http/https)."
                  }
                },
                "required": [
                  "url"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Session id from POST /v1/sessions."
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "404": {
            "description": "Structured error. `code` is stable and machine-readable; `message` states how to recover.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Structured error. `code` is stable and machine-readable; `message` states how to recover.",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Stable error code, e.g. \"stale_ref\", \"session_not_found\"."
                        },
                        "message": {
                          "type": "string",
                          "description": "Human/agent-readable explanation and recovery hint."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "502": {
            "description": "Structured error. `code` is stable and machine-readable; `message` states how to recover.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Structured error. `code` is stable and machine-readable; `message` states how to recover.",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Stable error code, e.g. \"stale_ref\", \"session_not_found\"."
                        },
                        "message": {
                          "type": "string",
                          "description": "Human/agent-readable explanation and recovery hint."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/sessions/{id}/back": {
      "post": {
        "summary": "Go back",
        "tags": [
          "Actions"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Session id from POST /v1/sessions."
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "404": {
            "description": "Structured error. `code` is stable and machine-readable; `message` states how to recover.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Structured error. `code` is stable and machine-readable; `message` states how to recover.",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Stable error code, e.g. \"stale_ref\", \"session_not_found\"."
                        },
                        "message": {
                          "type": "string",
                          "description": "Human/agent-readable explanation and recovery hint."
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": true
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/sessions/{id}/downloads/{downloadId}": {
      "get": {
        "summary": "Fetch a downloaded file",
        "tags": [
          "Observability"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "downloadId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
...
}